backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
onnxruntime
opencv-python-headless
exifread
google-cloud-tasks
google-cloud-storage
firebase-admin
python-multipart
numpy
scipy
redis
